FIX: Prevent removing cents

+++ dynamic_accounts_report/static/src/js/ageing.js
@@ -112,7 +112,7 @@ odoo.define('dynamic_accounts_report.ageing', function (require) {
                 if (typeof(amount) != 'number') {
                     amount = parseFloat(amount);
                 }
-                var formatted_value = (parseInt(amount)).toLocaleString(currency[2],{
+                var formatted_value = (parseFloat(amount)).toLocaleString(currency[2],{
                     minimumFractionDigits: 2
                 })
                 return formatted_value
