+++ dynamic_accounts_report/static/src/js/daybook.js
@@ -99,7 +99,7 @@ odoo.define('dynamic_partner_daybook.daybook', function (require) {
                 if (typeof(amount) != 'number') {
                     amount = parseFloat(amount);
                 }
-                var formatted_value = (parseInt(amount)).toLocaleString(currency[2],{
+                var formatted_value = (parseFloat(amount)).toLocaleString(currency[2],{
                     minimumFractionDigits: 2
                 })
                 return formatted_value
