+++ dynamic_accounts_report/static/src/js/general_ledger.js
@@ -220,7 +220,7 @@ odoo.define('dynamic_cash_flow_statements.general_ledger', function (require) {
                 if (typeof(amount) != 'number') {
                     amount = parseFloat(amount);
                 }
-                var formatted_value = (parseInt(amount)).toLocaleString(currency[2],{
+                var formatted_value = (parseFloat(amount)).toLocaleString(currency[2],{
                     minimumFractionDigits: 2
                 })
                 return formatted_value
